requirements.txt
tensorflow==2.18.0
Flask==3.0.0
pillow==10.2.0
numpy==1.26.3
h5py==3.12.1
